Documentation updated to match newer patterns
