--- StructuredData.swift
@@ -1,506 +1,0 @@
-import Foundation
-
-/// Represents structured data in JSON-LD format for rich snippets in search results.
-///
-/// The `StructuredData` struct provides a type-safe way to define structured data
-/// following various schema.org schemas like Article, Product, Organization, etc.
-public struct StructuredData {
-
-    /// The schema type for the structured data.
-    public enum SchemaType: String {
-        case article = "Article"
-        case blogPosting = "BlogPosting"
-        case breadcrumbList = "BreadcrumbList"
-        case course = "Course"
-        case event = "Event"
-        case faqPage = "FAQPage"
-        case howTo = "HowTo"
-        case localBusiness = "LocalBusiness"
-        case organization = "Organization"
-        case person = "Person"
-        case product = "Product"
-        case recipe = "Recipe"
-        case review = "Review"
-        case website = "WebSite"
-    }
-
-    /// The type of schema used for this structured data.
-    public let type: SchemaType
-
-    /// The raw data to be included in the structured data.
-    private let data: [String: Any]
-
-    /// Returns a copy of the raw data dictionary.
-    ///
-    /// - Returns: A dictionary containing the structured data properties.
-    public func getData() -> [String: Any] {
-        data
-    }
-
-    /// Creates structured data for an Article.
-    ///
-    /// - Parameters:
-    ///   - headline: The title of the article.
-    ///   - image: The URL to the featured image of the article.
-    ///   - author: The name or URL of the author.
-    ///   - publisher: Optional publisher, either as a StructuredData person or organization, or as a String name.
-    ///   - datePublished: The date the article was published.
-    ///   - dateModified: The date the article was last modified.
-    ///   - description: A short description of the article content.
-    ///   - url: The URL of the article.
-    /// - Returns: A structured data object for an article.
-    ///
-    /// - Example:
-    ///   ```swift
-    ///   // Using a String for publisher
-    ///   let articleData = StructuredData.article(
-    ///     headline: "How to Use WebUI",
-    ///     image: "https://example.com/images/article.jpg",
-    ///     author: "John Doe",
-    ///     publisher: "WebUI Blog",
-    ///     datePublished: Date(),
-    ///     description: "A guide to using WebUI for Swift developers"
-    ///   )
-    ///
-    ///   // Using a StructuredData organization as publisher
-    ///   let orgPublisher = StructuredData.organization(
-    ///     name: "WebUI Technologies",
-    ///     logo: "https://example.com/logo.png",
-    ///     url: "https://example.com"
-    ///   )
-    ///
-    ///   let articleWithOrg = StructuredData.article(
-    ///     headline: "How to Use WebUI",
-    ///     image: "https://example.com/images/article.jpg",
-    ///     author: "John Doe",
-    ///     publisher: orgPublisher,
-    ///     datePublished: Date(),
-    ///     description: "A guide to using WebUI for Swift developers"
-    ///   )
-    ///
-    ///   // Without a publisher
-    ///   let minimalArticle = StructuredData.article(
-    ///     headline: "Quick Tips",
-    ///     image: "https://example.com/images/tips.jpg",
-    ///     author: "Alex Developer",
-    ///     datePublished: Date()
-    ///   )
-    ///   ```
-    ///
-    /// - Note: For more control over the publisher entity, use the overloaded version
-    ///   of this method that accepts a StructuredData object as the publisher parameter.
-    public static func article(
-        headline: String,
-        image: String,
-        author: String,
-        publisher: Any? = nil,
-        datePublished: Date,
-        dateModified: Date? = nil,
-        description: String? = nil,
-        url: String? = nil
-    ) -> StructuredData {
-        var data: [String: Any] = [
-            "headline": headline,
-            "image": image,
-            "author": ["@type": "Person", "name": author],
-            "publisher": ["@type": "Organization", "name": publisher],
-            "datePublished": ISO8601DateFormatter().string(from: datePublished),
-        ]
-
-        // Handle different publisher types
-        if let publisher = publisher {
-            if let publisherName = publisher as? String {
-                data["publisher"] = ["@type": "Organization", "name": publisherName]
-            } else if let publisherData = publisher as? StructuredData {
-                if publisherData.type == .organization || publisherData.type == .person {
-                    // Extract the raw data from the structured data object
-                    let publisherDict = publisherData.getData()
-                    var typeDict = publisherDict
-                    typeDict["@type"] = publisherData.type.rawValue
-                    data["publisher"] = typeDict
-                }
-            }
-        }
-
-        if let dateModified = dateModified {
-            data["dateModified"] = ISO8601DateFormatter().string(from: dateModified)
-        }
-
-        if let description = description {
-            data["description"] = description
-        }
-
-        if let url = url {
-            data["url"] = url
-        }
-
-        return StructuredData(type: .article, data: data)
-    }
-
-    /// Creates structured data for a product.
-    ///
-    /// - Parameters:
-    ///   - name: The name of the product.
-    ///   - image: The URL to the product image.
-    ///   - description: A description of the product.
-    ///   - sku: The Stock Keeping Unit identifier.
-    ///   - brand: The brand name of the product.
-    ///   - offers: The offer details (price, availability, etc.).
-    ///   - review: Optional review information.
-    /// - Returns: A structured data object for a product.
-    ///
-    /// - Example:
-    ///   ```swift
-    ///   let productData = StructuredData.product(
-    ///     name: "Swift WebUI Course",
-    ///     image: "https://example.com/images/course.jpg",
-    ///     description: "Master WebUI development with Swift",
-    ///     sku: "WEBUI-101",
-    ///     brand: "Swift Academy",
-    ///     offers: ["price": "99.99", "priceCurrency": "USD", "availability": "InStock"]
-    ///   )
-    ///   ```
-    public static func product(
-        name: String,
-        image: String,
-        description: String,
-        sku: String,
-        brand: String,
-        offers: [String: Any],
-        review: [String: Any]? = nil
-    ) -> StructuredData {
-        var data: [String: Any] = [
-            "name": name,
-            "image": image,
-            "description": description,
-            "sku": sku,
-            "brand": ["@type": "Brand", "name": brand],
-            "offers": offers.merging(["@type": "Offer"]) { _, new in new },
-        ]
-
-        if let review = review {
-            data["review"] = review.merging(["@type": "Review"]) { _, new in new }
-        }
-
-        return StructuredData(type: .product, data: data)
-    }
-
-    /// Creates structured data for an organization.
-    ///
-    /// - Parameters:
-    ///   - name: The name of the organization.
-    ///   - logo: The URL to the organization's logo.
-    ///   - url: The URL of the organization's website.
-    ///   - contactPoint: Optional contact information.
-    ///   - sameAs: Optional array of URLs that also represent the entity.
-    /// - Returns: A structured data object for an organization.
-    ///
-    /// - Example:
-    ///   ```swift
-    ///   let orgData = StructuredData.organization(
-    ///     name: "WebUI Technologies",
-    ///     logo: "https://example.com/logo.png",
-    ///     url: "https://example.com",
-    ///     sameAs: ["https://twitter.com/webui", "https://github.com/webui"]
-    ///   )
-    ///   ```
-    public static func organization(
-        name: String,
-        logo: String,
-        url: String,
-        contactPoint: [String: Any]? = nil,
-        sameAs: [String]? = nil
-    ) -> StructuredData {
-        var data: [String: Any] = [
-            "name": name,
-            "logo": logo,
-            "url": url,
-        ]
-
-        if let contactPoint = contactPoint {
-            data["contactPoint"] = contactPoint.merging(["@type": "ContactPoint"]) { _, new in new }
-        }
-
-        if let sameAs = sameAs {
-            data["sameAs"] = sameAs
-        }
-
-        return StructuredData(type: .organization, data: data)
-    }
-
-    /// Creates structured data for a person.
-    ///
-    /// - Parameters:
-    ///   - name: The name of the person.
-    ///   - givenName: The given (first) name of the person.
-    ///   - familyName: The family (last) name of the person.
-    ///   - image: The URL to an image of the person.
-    ///   - jobTitle: The person's job title.
-    ///   - email: The person's email address.
-    ///   - telephone: The person's telephone number.
-    ///   - url: The URL of the person's website or profile.
-    ///   - address: Optional address information.
-    ///   - birthDate: The person's date of birth.
-    ///   - sameAs: Optional array of URLs that also represent the person (social profiles).
-    /// - Returns: A structured data object for a person.
-    ///
-    /// - Example:
-    ///   ```swift
-    ///   let personData = StructuredData.person(
-    ///     name: "Jane Doe",
-    ///     givenName: "Jane",
-    ///     familyName: "Doe",
-    ///     jobTitle: "Software Engineer",
-    ///     url: "https://janedoe.com",
-    ///     sameAs: ["https://twitter.com/janedoe", "https://github.com/janedoe"]
-    ///   )
-    ///   ```
-    public static func person(
-        name: String,
-        givenName: String? = nil,
-        familyName: String? = nil,
-        image: String? = nil,
-        jobTitle: String? = nil,
-        email: String? = nil,
-        telephone: String? = nil,
-        url: String? = nil,
-        address: [String: Any]? = nil,
-        birthDate: Date? = nil,
-        sameAs: [String]? = nil
-    ) -> StructuredData {
-        var data: [String: Any] = [
-            "name": name
-        ]
-
-        if let givenName = givenName {
-            data["givenName"] = givenName
-        }
-
-        if let familyName = familyName {
-            data["familyName"] = familyName
-        }
-
-        if let image = image {
-            data["image"] = image
-        }
-
-        if let jobTitle = jobTitle {
-            data["jobTitle"] = jobTitle
-        }
-
-        if let email = email {
-            data["email"] = email
-        }
-
-        if let telephone = telephone {
-            data["telephone"] = telephone
-        }
-
-        if let url = url {
-            data["url"] = url
-        }
-
-        if let address = address {
-            data["address"] = address.merging(["@type": "PostalAddress"]) { _, new in new }
-        }
-
-        if let birthDate = birthDate {
-            data["birthDate"] = ISO8601DateFormatter().string(from: birthDate)
-        }
-
-        if let sameAs = sameAs {
-            data["sameAs"] = sameAs
-        }
-
-        return StructuredData(type: .person, data: data)
-    }
-
-    /// Creates structured data for breadcrumbs navigation.
-    ///
-    /// - Parameter items: Array of breadcrumb items with name, item (URL), and position.
-    /// - Returns: A structured data object for breadcrumbs navigation.
-    ///
-    /// - Example:
-    ///   ```swift
-    ///   let breadcrumbsData = StructuredData.breadcrumbs([
-    ///     ["name": "Home", "item": "https://example.com", "position": 1],
-    ///     ["name": "Blog", "item": "https://example.com/blog", "position": 2],
-    ///     ["name": "Article Title", "item": "https://example.com/blog/article", "position": 3]
-    ///   ])
-    ///   ```
-    public static func breadcrumbs(_ items: [[String: Any]]) -> StructuredData {
-        let itemListElements = items.map { item in
-            var element: [String: Any] = ["@type": "ListItem"]
-
-            if let name = item["name"] as? String {
-                element["name"] = name
-            }
-
-            if let itemUrl = item["item"] as? String {
-                element["item"] = itemUrl
-            }
-
-            if let position = item["position"] as? Int {
-                element["position"] = position
-            }
-
-            return element
-        }
-
-        return StructuredData(type: .breadcrumbList, data: ["itemListElement": itemListElements])
-    }
-
-    /// Creates structured data for an Article with a StructuredData publisher.
-    ///
-    /// - Parameters:
-    ///   - headline: The title of the article.
-    ///   - image: The URL to the featured image of the article.
-    ///   - author: The name or URL of the author.
-    ///   - publisher: Optional StructuredData object representing the publisher as a person or organization.
-    ///   - datePublished: The date the article was published.
-    ///   - dateModified: The date the article was last modified.
-    ///   - description: A short description of the article content.
-    ///   - url: The URL of the article.
-    /// - Returns: A structured data object for an article.
-    ///
-    /// - Example:
-    ///   ```swift
-    ///   // Using an organization as publisher
-    ///   let organization = StructuredData.organization(
-    ///     name: "WebUI Technologies",
-    ///     logo: "https://example.com/logo.png",
-    ///     url: "https://example.com"
-    ///   )
-    ///
-    ///   let articleWithOrg = StructuredData.article(
-    ///     headline: "How to Use WebUI",
-    ///     image: "https://example.com/images/article.jpg",
-    ///     author: "John Doe",
-    ///     publisher: organization,
-    ///     datePublished: Date(),
-    ///     description: "A guide to using WebUI for Swift developers"
-    ///   )
-    ///
-    ///   // Using a person as publisher
-    ///   let personPublisher = StructuredData.person(
-    ///     name: "Jane Doe",
-    ///     url: "https://janedoe.com"
-    ///   )
-    ///
-    ///   let articleWithPerson = StructuredData.article(
-    ///     headline: "My WebUI Journey",
-    ///     image: "https://example.com/images/journey.jpg",
-    ///     author: "John Smith",
-    ///     publisher: personPublisher,
-    ///     datePublished: Date(),
-    ///     description: "Personal experiences with WebUI framework"
-    ///   )
-    ///
-    ///   // Without a publisher
-    ///   let minimalArticle = StructuredData.article(
-    ///     headline: "Quick Tips",
-    ///     image: "https://example.com/images/tips.jpg",
-    ///     author: "Alex Developer",
-    ///     datePublished: Date()
-    ///   )
-    ///   ```
-    public static func article(
-        headline: String,
-        image: String,
-        author: String,
-        publisher: StructuredData?,
-        datePublished: Date,
-        dateModified: Date? = nil,
-        description: String? = nil,
-        url: String? = nil
-    ) -> StructuredData {
-        var data: [String: Any] = [
-            "headline": headline,
-            "image": image,
-            "author": ["@type": "Person", "name": author],
-            "datePublished": ISO8601DateFormatter().string(from: datePublished),
-        ]
-
-        if let publisher = publisher {
-            if publisher.type == .organization || publisher.type == .person {
-                // Extract the raw data from the structured data object
-                let publisherDict = publisher.getData()
-                var typeDict = publisherDict
-                typeDict["@type"] = publisher.type.rawValue
-                data["publisher"] = typeDict
-            }
-        }
-
-        if let dateModified = dateModified {
-            data["dateModified"] = ISO8601DateFormatter().string(from: dateModified)
-        }
-
-        if let description = description {
-            data["description"] = description
-        }
-
-        if let url = url {
-            data["url"] = url
-        }
-
-        return StructuredData(type: .article, data: data)
-    }
-
-    /// Creates a custom structured data object with the specified schema type and data.
-    ///
-    /// - Parameters:
-    ///   - type: The schema type for the structured data.
-    ///   - data: The data to include in the structured data.
-    /// - Returns: A structured data object with the specified type and data.
-    ///
-    /// - Example:
-    ///   ```swift
-    ///   let customData = StructuredData.custom(
-    ///     type: .review,
-    ///     data: [
-    ///       "itemReviewed": ["@type": "Product", "name": "WebUI Framework"],
-    ///       "reviewRating": ["@type": "Rating", "ratingValue": "5"],
-    ///       "author": ["@type": "Person", "name": "Jane Developer"]
-    ///     ]
-    ///   )
-    ///   ```
-    public static func custom(type: SchemaType, data: [String: Any]) -> StructuredData {
-        StructuredData(type: type, data: data)
-    }
-
-    /// Initializes a new structured data object with the specified schema type and data.
-    ///
-    /// - Parameters:
-    ///   - type: The schema type for the structured data.
-    ///   - data: The data to include in the structured data.
-    public init(type: SchemaType, data: [String: Any]) {
-        self.type = type
-        self.data = data
-    }
-
-    /// Converts the structured data to a JSON string.
-    ///
-    /// - Returns: A JSON string representation of the structured data, or an empty string if serialization fails.
-    public func toJSON() -> String {
-        var jsonObject: [String: Any] = [
-            "@context": "https://schema.org",
-            "@type": type.rawValue,
-        ]
-
-        // Merge the data dictionary with the base JSON object
-        for (key, value) in data {
-            jsonObject[key] = value
-        }
-
-        // Try to serialize the JSON object to data
-        if let jsonData = try? JSONSerialization.data(
-            withJSONObject: jsonObject,
-            options: [.prettyPrinted, .withoutEscapingSlashes]
-        ) {
-            // Convert the data to a string
-            return String(data: jsonData, encoding: .utf8) ?? ""
-        }
-
-        return ""
-    }
-}
